Removed complexity in ResizeableMaze, by removing component state.
Before, ResizeableMaze was keeping track of components created from TileMaker. It now asks for new components each render. Entities required memoization to avoid flickering--we can expect similar changes to be made when we implement images to tiles.

Changing state in ResizeableMaze no longer calls render. It is up to the user to detect changes and then render, or more likely, to call render periodically.